Moved exception into appropriate packages
