--- webtester-support-junit5/src/main/java/info/novatec/testit/webtester/junit5/exceptions/NonUniqueBrowserNameException.java
@@ -1,18 +1,0 @@
-package info.novatec.testit.webtester.junit5.exceptions;
-
-import info.novatec.testit.webtester.browser.Browser;
-
-
-/**
- * This exception is thrown in case there is more than one managed {@link Browser} with the same name when execution an
- * extension.
- *
- * @since 2.1
- */
-public class NonUniqueBrowserNameException extends TestClassFormatException {
-
-    public NonUniqueBrowserNameException() {
-        super("every browser needs a unique name!");
-    }
-
-}
